requirements.txt
flask==3.0.0
line-bot-sdk==3.5.0
python-dotenv==1.0.0
sentry-sdk[flask]==1.38.0
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.17.0
google-generativeai==0.3.1
gunicorn==21.2.0
redis==5.0.1
neo4j==5.15.0
jieba==0.42.1